Support uint64 short URL sequence and extend Base58 ID format
Migrate counter interfaces and storage from uint32 to uint64, update Feistel ID generation to preserve 6-char compact IDs while appending a Base58 suffix as needed, extend valid short ID length to 12, and document the format in README.

+++ internal/service/id_counter.go
@@ -8,5 +8,5 @@ import (
 
 // ShortURLCounter describes counter operations required by ShortURLService.
 type ShortURLCounter interface {
-	Next(ctx context.Context) (uint32, error)
+	Next(ctx context.Context) (uint64, error)
 }
